third_party/WebKit/LayoutTests/http/tests/inspector/network/network-xhr-sync.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    InspectorTest.recordNetwork();
    InspectorTest.makeSimpleXHR("GET", "resources/resource.php", false, step2);

    function step2()
    {
        var request1 = InspectorTest.networkRequests().pop();
        InspectorTest.addResult(request1.url());
        InspectorTest.addResult("resource.type: " + request1.resourceType());
        InspectorTest.assertTrue(!request1.failed, "Resource loading failed.");
        request1.requestContent().then(step3);
    }

    function step3(content)
    {
        InspectorTest.addResult("resource.content after requesting content: " + content);

        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
 <p>
 Tests XHR network resource type and content for synchronous requests.
<a href="https://bugs.webkit.org/show_bug.cgi?id=61205">Bug 61205</a>
</p>
</body>
</html>
