third_party/WebKit/LayoutTests/http/tests/inspector/network/network-xhr-sync-expected.txt

CONSOLE WARNING: line 33: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests XHR network resource type and content for synchronous requests. Bug 61205

http://127.0.0.1:8000/inspector/network/resources/resource.php
resource.type: xhr
resource.content after requesting content: Hello world
